changed initSession, changed client and page, change db
