+++ src/database/db.ts
@@ -118,6 +118,10 @@ export async function GetSessionData(sid : string) : Promise<any> {
         throw new Error("Wrong guest code")
     }
 
+    if(hostId[0].host_id === null){ //here because the queue mount runs before the host is set
+        throw new Error("null host_id")
+    }
+
     const hostName : any[] = await sql`
         SELECT username FROM users
         WHERE user_id = ${hostId[0].host_id}
@@ -128,14 +132,13 @@ export async function GetSessionData(sid : string) : Promise<any> {
         WHERE session_id = ${sid} AND user_id <> ${hostId[0].host_id}
     `
 
-    // Returns queue in sorted order
+    //Returns queue in sorted order
     const queue : any[] = await sql`
-        SELECT song_name, artist_name, album_cover, placement
+        SELECT song_id, song_name, artist_name, album_cover, placement
         FROM queues
         WHERE session_id = ${sid}
         ORDER BY placement
     `
-    
     return {hostName: hostName[0].username, clientNames: clientNames, queue: queue}; // Return an object containing the hosts user name
 }
